Payload Problem Resolved

+++ components/org.apache.stratos.autoscaler/src/main/java/org/apache/stratos/autoscaler/context/cluster/ClusterInstanceContext.java
@@ -582,13 +582,13 @@ public class ClusterInstanceContext extends InstanceContext {
         this.dependentScaleCheckFactHandle = dependentScaleCheckFactHandle;
     }
 
-    public void setCurveFinderCoefficientsOfMemoryConsumption(float a, float b, float c){
+    public void setCurveFinderCoefficientsOfMemoryConsumption(double a, double b, double c){
         memoryConsumption.setA(a);
         memoryConsumption.setB(b);
         memoryConsumption.setC(c);
     }
 
-    public void setCurveFinderCoefficientsOfLoadAverage(float a, float b, float c){
+    public void setCurveFinderCoefficientsOfLoadAverage(double a, double b, double c){
         loadAverage.setA(a);
         loadAverage.setB(b);
         loadAverage.setC(c);
