+++ components/org.apache.stratos.messaging/src/main/java/org/apache/stratos/messaging/event/health/stat/CurveFinderOfLoadAverageEvent.java
@@ -29,9 +29,9 @@ public class CurveFinderOfLoadAverageEvent extends Event {
     /**
      * Curve finder co-efficients
      */
-    private final float a;
-    private final float b;
-    private final float c;
+    private final double a;
+    private final double b;
+    private final double c;
 
     public CurveFinderOfLoadAverageEvent(String networkPartitionId, String clusterId, String clusterInstanceId, double value, float a, float b, float c) {
         this.networkPartitionId = networkPartitionId;
@@ -60,15 +60,15 @@ public class CurveFinderOfLoadAverageEvent extends Event {
         return value;
     }
 
-    public float getA() {
+    public double getA() {
         return a;
     }
 
-    public float getB() {
+    public double getB() {
         return b;
     }
 
-    public float getC() {
+    public double getC() {
         return c;
     }
 }
