+++ components/org.apache.stratos.messaging/src/main/java/org/apache/stratos/messaging/event/health/stat/MemberCurveFinderOfLoadAverageEvent.java
@@ -13,12 +13,12 @@ public class MemberCurveFinderOfLoadAverageEvent extends Event {
     /**
      * Curve finder co-efficients
      */
-    private final float a;
-    private final float b;
-    private final float c;
+    private final double a;
+    private final double b;
+    private final double c;
 
 
-    public MemberCurveFinderOfLoadAverageEvent(String clusterId, String clusterInstanceId, String memberId, double value, float a, float b, float c) {
+    public MemberCurveFinderOfLoadAverageEvent(String clusterId, String clusterInstanceId, String memberId, double value, double a, double b, double c) {
         this.clusterId = clusterId;
         this.clusterInstanceId = clusterInstanceId;
         this.memberId = memberId;
@@ -45,15 +45,15 @@ public class MemberCurveFinderOfLoadAverageEvent extends Event {
         return value;
     }
 
-    public float getA() {
+    public double getA() {
         return a;
     }
 
-    public float getB() {
+    public double getB() {
         return b;
     }
 
-    public float getC() {
+    public double getC() {
         return c;
     }
 }
